move cartitem object to domain layer
